finished migrating all the existing cmtc commands

+++ cmtc/src/main/scala/cmt/client/command/GotoFirstExercise.scala
@@ -28,7 +28,7 @@ object GotoFirstExercise:
       def execute(): Either[CmtError, String] = {
         val config = new CMTcConfig(options.shared.studentifiedRepo.value)
         GotoExercise
-          .Options(exercise = ExerciseId(config.exercises.head.toInt), force = options.force, shared = options.shared)
+          .Options(exercise = ExerciseId(config.exercises.head), force = options.force, shared = options.shared)
           .execute()
       }
