+++ cmtc/src/main/scala/cmt/client/command/execution/GoToFirstExercise.scala
@@ -21,8 +21,5 @@ import cmt.core.execution.Executable
 given Executable[GotoFirstExercise] with
   extension (cmd: GotoFirstExercise)
     def execute(): Either[CmtError, String] =
-      GotoExercise(
-        cmd.config,
-        cmd.forceMoveToExercise,
-        cmd.studentifiedRepo,
-        ExerciseId(cmd.config.exercises.head.toInt)).execute()
+      GotoExercise(cmd.config, cmd.forceMoveToExercise, cmd.studentifiedRepo, ExerciseId(cmd.config.exercises.head))
+        .execute()
